third_party/blink/web_tests/platform/generic/fast/loader/show-only-one-beforeunload-dialog-expected.txt

CONFIRM NAVIGATION
CONSOLE ERROR: Blocked attempt to show multiple 'beforeunload' confirmation panels for a single navigation.
CONSOLE ERROR: Blocked attempt to show multiple 'beforeunload' confirmation panels for a single navigation.
